(ref) integrate setup_tools package
